update work repository and work serviceimpl

+++ src/main/java/com/laura/carpaciu/dao/interfaces/WorkRepository.java
@@ -16,7 +16,7 @@ public interface WorkRepository extends CrudRepository<Work, Long> {
 
 	Optional<Work> findById(int id);
 
-	int updateWorkTimeAndDescription(double timedWork, String workDescription, int id);
+	int updateWorkTimeAndDescription(double timedWork, String workDescription, Long id);
 
 	int deleteWork(int id);
